agentes/05_account_manager/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
anthropic>=0.40
requests>=2.31
python-dotenv>=1.0
pydantic>=2.6
